Do not activate context
After a setting is saved, do not activate the settings. Let the user
activate the context, as it may incure extra cost. If the context
is active already, give the user a warning that this action will
deactive it. Reset  Username and Password for Plans which do not
contain these settings.